window: frame clauses, NULLS ordering, reducer-args, dialect aliases
Burn down window-HIR/lowering/dialect gaps discovered while auditing
WindowFunctionTest and ExtendWindowCheckerTest. All 25/25 WindowFunctionTest
and 120/120 ExtendWindowCheckerTest now pass. Net engine suite: 2669/277/612/19
(was 299/615); +22 tests passing, zero regressions.

Frame clauses (ROWS/RANGE BETWEEN ... AND ...).
  SqlExpr gains WindowFrame, FrameType, and a sealed FrameBound hierarchy
  (Unbounded/CurrentRow/Offset) that each render themselves contextually
  based on start/end position. SqlExpr.WindowCall grows an Optional<WindowFrame>
  field. ExtendLowering.lowerFrame structurally translates TypedFrame to
  the SQL-side record.

NULLS FIRST/LAST on ORDER BY.
  ExtendLowering.lowerSortKey now emits NULLS FIRST for DESC and NULLS LAST
  for ASC, matching legacy PlanGenerator semantics and Pure/Legend convention.

Decimal RANGE bounds (e.g. 0.5->_range(2.5)).
  Widen Offset.value from long to double end-to-end - Offset,
  SqlExpr.OffsetFrameBound, ExtendChecker.extractNumericValue - with integral
  values still rendering without a decimal point.

Reducer extra args (joinStrings separator).
  Aggregate windows like {p,w,r|$r.name}:y|$y->joinStrings('') were dropping
  the separator literal. New extractReducerExtraArgs walks the reducer body
  and appends non-variable args to funcArgs so STRING_AGG(name, '') renders.

DuckDBDialect aliases (legacy mapPureFuncToSql compatibility).
  Bare Pure-native names now map to their SQL counterparts:
    first/firstValue -> FIRST_VALUE, last/lastValue -> LAST_VALUE,
    nth/nthValue    -> NTH_VALUE,
    percentile/percentileCont -> QUANTILE_CONT (bare percentile = continuous).

Known follow-ups (filed for separate PRs, not bundled):
  - TypeInfo is entirely dead post-HIR port; purge in a dedicated commit.
  - Three parallel FrameType enums (parser / HIR / sqlgen) could collapse to one.
  - TypedFrameBound and SqlExpr.FrameBound are structurally identical parallel
    hierarchies; merging would cross the compiler/sqlgen layer seam.
  - OrderByTerm.direction is stringly-typed; promote to enum.

+++ engine/src/main/java/com/gs/legend/compiler/typed/Offset.java
@@ -1,4 +1,12 @@
 package com.gs.legend.compiler.typed;
 
-/** A row/range offset as a window frame bound. */
-public record Offset(long value) implements TypedFrameBound {}
+/**
+ * A row/range offset as a window frame bound.
+ *
+ * <p>Value type is {@code double} to support fractional RANGE bounds
+ * (e.g. {@code 0.5->_range(2.5)}). ROWS offsets are naturally integral
+ * but share the representation; integral values round-trip exactly.
+ * Sign convention: positive = FOLLOWING, negative = PRECEDING, zero =
+ * CURRENT ROW (normalized by the checker to the {@link CurrentRow} bound).
+ */
+public record Offset(double value) implements TypedFrameBound {}
